chore: migrated customerio to common (#411)
