+++ Sources/Tracking/autogenerated/AutoMockable.generated.swift
@@ -81,49 +81,6 @@ import CioInternalCommon
  */
 
 /**
- Class to easily create a mocked version of the `CleanupRepository` class.
- This class is equipped with functions and properties ready for you to mock!
-
- Note: This file is automatically generated. This means the mocks should always be up-to-date and has a consistent API.
- See the SDK documentation to learn the basics behind using the mock classes in the SDK.
- */
-class CleanupRepositoryMock: CleanupRepository, Mock {
-    /// If *any* interactions done on mock. `true` if any method or property getter/setter called.
-    var mockCalled: Bool = false //
-
-    init() {
-        Mocks.shared.add(mock: self)
-    }
-
-    public func resetMock() {
-        cleanupCallsCount = 0
-
-        mockCalled = false // do last as resetting properties above can make this true
-    }
-
-    // MARK: - cleanup
-
-    /// Number of times the function was called.
-    private(set) var cleanupCallsCount = 0
-    /// `true` if the function was ever called.
-    var cleanupCalled: Bool {
-        cleanupCallsCount > 0
-    }
-
-    /**
-     Set closure to get called when function gets called. Great way to test logic or return a value for the function.
-     */
-    var cleanupClosure: (() -> Void)?
-
-    /// Mocked function for `cleanup()`. Your opportunity to return a mocked value and check result of mock in test code.
-    func cleanup() {
-        mockCalled = true
-        cleanupCallsCount += 1
-        cleanupClosure?()
-    }
-}
-
-/**
  Class to easily create a mocked version of the `DeviceAttributesProvider` class.
  This class is equipped with functions and properties ready for you to mock!
